% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{do.ralget}
\alias{do.ralget}
\title{Set value for variable in DAG}
\usage{
\method{do}{ralget}(graph, ...)
}
\arguments{
\item{graph}{ralget}

\item{...}{assignment expressions for node in DAG}
}
\description{
Set value for variable in DAG
}
