% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{manipulate}
\alias{manipulate}
\title{Set value for variable in DAG}
\usage{
manipulate(graph, ...)
}
\arguments{
\item{graph}{ralget}

\item{...}{assignment expressions for node in DAG}
}
\description{
Set value for variable in DAG
}
