% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ralget}
\alias{plot.ralget}
\title{Set value for variable in DAG}
\usage{
\method{plot}{ralget}(x, ...)
}
\arguments{
\item{x}{ralget}

\item{...}{other args}
}
\value{
ralget
}
\description{
Set value for variable in DAG
}
